plugin/channel/discord_bot/requirements.txt
# Dipendenze del plugin discord_bot.
# Ogni plugin deve dichiarare TUTTE le sue dipendenze, anche se già presenti nel progetto.
# Il loader le installa/verifica automaticamente all'avvio.
agno>=2.5.11
discord.py>=2.3.0
pyyaml>=6.0.3
